Rename folders back to original names
