+++ src/main/java/frc/robot/subsystems/drivetrain/module/ModuleSim.java
@@ -1,4 +1,4 @@
-package frc.robot.subsystems.drivetrain2.module2;
+package frc.robot.subsystems.drivetrain.module;
 
 import com.ctre.phoenix.motorcontrol.can.WPI_TalonFX;
